+++ Package.swift
@@ -12,7 +12,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/wultra/powerauth-mobile-sdk-spm.git", .upToNextMinor(from: "1.9.2")),
-        .package(url: "https://github.com/wultra/networking-apple.git", .upToNextMinor(from: "1.5.0"))
+        .package(url: "https://github.com/wultra/networking-apple.git", .upToNextMinor(from: "1.5.1"))
     ],
     targets: [
         .target(